Fix missing os import and change precision from float32 to float64

+++ preprocessing/preprocessing/utils/utils.py
@@ -80,10 +80,10 @@ def read_point_cloud(point_cloud_file):
 
     if ext == '.bin':
         point_cloud = (np.fromfile(point_cloud_file,
-                                   dtype=np.float32).reshape(-1, 6))
+                                   dtype=np.float64).reshape(-1, 6))
     elif ext == '.npy':
         point_cloud = (np.load(point_cloud_file,
-                               dtype=np.float32).reshape(-1, 6))
+                               dtype=np.float64).reshape(-1, 6))
     else:
         print("Invalid point cloud format encountered.")
         sys.exit()
